requirements-dev.txt
pytest==3.8.2
pytest-cov==2.6.0
pytest-django==3.4.3
flake8==3.5.0
responses==0.10.6
black==18.9b0
isort==4.3.4
pre-commit==1.21.0
mock
